refactor: modularize PGLite storage adapter for improved maintainability
Phase 1 of large file refactoring project (Issue #228)

## Changes Made

### 🔧 Separated Components
- **PGlite Dialect**: Moved to `src/storage/dialects/pglite-dialect.ts` (117 lines)
  - PGliteConnection, PGliteDriver, PGliteDialect classes
  - PostgreSQL compatibility layer for PGlite WASM

- **Database Errors**: Moved to `src/storage/errors/database-error.ts` (36 lines)
  - DatabaseError class with error codes
  - Test tracking utilities

- **Kysely Types**: Moved to `src/storage/types/kysely-types.ts` (60 lines)
  - CallEdgeTable, InternalCallEdgeTable interfaces
  - Database schema definitions

### 📊 Impact
- **Main file reduced**: 5,706 → 5,537 lines (169 lines extracted)
- **New modular structure**: 4 focused files with clear responsibilities
- **Type safety maintained**: All imports/exports properly structured
- **Backward compatibility**: Re-exports maintain existing API

### ✅ Quality Improvements
- **Separation of concerns**: Each module has single responsibility
- **Improved testability**: Components can be tested in isolation
- **Enhanced maintainability**: Smaller, focused files easier to understand
- **TypeScript compilation**: All type errors resolved ✓

## Next Steps
- Phase 2: Further break down remaining 5,537 lines into feature-specific modules
- Phase 3: Apply same approach to other large files (staged-analysis-engine.ts, cli.ts)
